Update package versions and dependencies
The commit includes changes to update the versions of several packages, as well as adjustments to some of their dependencies. It involves packages such as amphp/amp, amphp/byte-stream, composer/pcre, and several others. This update ensures that the application uses the latest, most secure, and most efficient versions of these packages.