Add IndexPageElements class for managing page content
Created a new file called IndexPageElements to store all elements from the configuration that are used in the IndexPage. This simplifies the process of getting the data and allows for more efficient handling of the elements within the IndexPage class. The changes make code more readable and easier to manage.